Fix Gradle detekt issues
