+++ app/src/main/kotlin/br/com/colman/petals/components/DateTimePicker.kt
@@ -23,7 +23,7 @@ fun dateDialogState(onDateChange: (LocalDate) -> Unit) = createMaterialDialog {
 
 @Composable
 fun timeDialogState(is24Hour: Boolean, onTimeChange: (newTime: LocalTime) -> Unit) = createMaterialDialog {
-  timepicker(title = stringResource(select_time)) { time ->
+  timepicker(title = stringResource(select_time), is24HourClock = is24Hour) { time ->
     onTimeChange(time)
   }
 }
